Merged PR 1559: Set isNotificationOnly to true to prevent users from replying
Set isNotificationOnly to true to prevent users from replying to the messages sent by Company Communicator.

This was originally set to false to allow for recovery in case CC missed the conversationUpdate, but now that we can get the chat thread ID via Graph we can set this to true.