Merged PR 2183: [ccv4.0.1]Refactoring export function and update to azure storage blob v12
This PR includes refactoring export function and update to azure storage blob v12.

Changes Included :
1)[Refactoring Export function 1448291](https://domoreexp.visualstudio.com/MSTeams/_workitems/edit/1448291)

1) ExportOrchestration - Removed the dependencies from the orchestrator and invoked the activity directly.
2) Activity classes - Refactored the activity classes to only have activity function and removed the earlier pattern of having another function using context to call the activity.
3) Refactored classes to remove concrete dependency and make them unit testable.

2)[Upload Azure storage blob v12 1465479](https://domoreexp.visualstudio.com/MSTeams/_workitems/edit/1465479)

Updated the Azure storage package from v11 to v12. As part of this rewritten the upload function in Upload Activity using v12 package. The upload strategy is same as before and no change to the functionality.

3)Updated unit test

Verification:
Pending for large data.