Merged PR 2512: [CC] "Duplicate" option leads to a state where the "..." menu shows then disappears immediately
This PR addresses the duplicate issue which leads the menu state to be unpredictable.

Issue :
Earlier the menuOpen state(of Menu Component) was being externally managed using the event onMenuOpenChange and local state "menuOpen". Since the state was being updated using setState, the asynchronous behavior was leading it to being unpredictable.
Reasons to use external state :
- immediately switching off menu on click of submenu item and let the async call inside the onclick method not block the UI.
- managing active state.

Fix :
Reuse internal state of the component and use external state only to make the state as false immediately  on click of submenu item.