Scale for Large audience and theme fixes. (#368)
* Merged PR 2636: [CCv4.1] Fix task module theme

This PR fixes the issue with themes in task module pages.

Issue :
On change of the theme of Teams App from Default to Dark or High Contrast. When a user clicks on Task Module( any popup pages ), theme is not applied.

Fix:
This is a regression issue.
In 4.1 white color was added as a background in the leaf level which was overriding the theme effects.
The fix was to remove the white color from the leaf level.

Verification : Done

* Merged PR 2664: [CC] Scale for larger audience

The GetRecipientsActivity is changed to handle paged db data and will return token if more data is available.
The activity response is capped at 100000 users. This will ensure we don't run into serialization issues and can handle large audiences.

Co-authored-by: Priyank Saxena <prsaxen@microsoft.com>