Fix PowerShell deployment script with targetting
Due to the addition of the TargetingEnabled and MasterAdminUpns parameters to azuredeploy.json the PowerShell script for deployment stopped working.

This was due to the fact that azuredeploy defined a min length of 1 for MasterAdminUpns while the deploy.ps1 script was not taking into account that new parameter, preventing the user from being able to provide that value.

With this fix, the user will now be able to provide said parameter.