Update deploy.ps1
A couple of minor best practices changes around null checking and using aliases to maintain readability of the code.

Azure CLI tools will now download the latest version if nothing is installed, as well as explicitly use msiexec from system and not rely on path for that.

Also, allowing AzureADPreview to satisfy the requirement for AzureAD being installed. I left it as a warning in the messaging to the user.